# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Local-First Collaborative Document Editor API",
    "description": "API documentation for managing documents, workspace sharing, audit logging, snapshots, and offline operational sync. Enforces cookie-based session security via NextAuth.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/documents/{id}": {
      "get": {
        "summary": "Get Document Details",
        "description": "Securely retrieves details of a specific document.",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved document detail.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid", "example": "77777777-7777-4777-8777-777777777777" },
                        "title": { "type": "string", "example": "My Design Doc" },
                        "ownerId": { "type": "string", "format": "uuid", "example": "e8a1f802-95bd-44ef-ad79-c5c2494537dc" },
                        "createdAt": { "type": "string", "format": "date-time" },
                        "updatedAt": { "type": "string", "format": "date-time" },
                        "role": { "type": "string", "enum": ["OWNER", "EDITOR", "VIEWER"], "example": "OWNER" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "403": { "$ref": "#/components/responses/403Forbidden" },
          "404": { "$ref": "#/components/responses/404NotFound" }
        }
      },
      "patch": {
        "summary": "Rename Document",
        "description": "Securely renames/updates document title.",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 100, "example": "Renamed Collaborative Doc" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document renamed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid", "example": "77777777-7777-4777-8777-777777777777" },
                        "title": { "type": "string", "example": "Renamed Collaborative Doc" },
                        "updatedAt": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "403": { "$ref": "#/components/responses/403Forbidden" },
          "404": { "$ref": "#/components/responses/404NotFound" },
          "422": { "$ref": "#/components/responses/422ValidationError" }
        }
      },
      "delete": {
        "summary": "Delete Document",
        "description": "Securely deletes a specific document. Owner-only.",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "responses": {
          "200": {
            "description": "Document deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Document successfully deleted." }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "403": { "$ref": "#/components/responses/403Forbidden" },
          "404": { "$ref": "#/components/responses/404NotFound" }
        }
      }
    },
...
    "/api/documents/{id}/member": {
      "patch": {
        "summary": "Change Member Role",
        "description": "Updates a document member's role. Owner-only.",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["memberId", "role"],
                "properties": {
                  "memberId": { "type": "string", "format": "uuid", "example": "11111111-1111-4111-8111-111111111111" },
                  "role": { "type": "string", "enum": ["OWNER", "EDITOR", "VIEWER"], "example": "VIEWER" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collaborator role successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Collaborator role successfully updated." },
                    "membership": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid", "example": "11111111-1111-4111-8111-111111111111" },
                        "documentId": { "type": "string", "format": "uuid", "example": "77777777-7777-4777-8777-777777777777" },
                        "userId": { "type": "string", "format": "uuid", "example": "e8a1f802-95bd-44ef-ad79-c5c2494537dc" },
                        "role": { "type": "string", "enum": ["OWNER", "EDITOR", "VIEWER"], "example": "VIEWER" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "403": { "$ref": "#/components/responses/403Forbidden" },
          "422": { "$ref": "#/components/responses/422ValidationError" }
        }
      },
      "delete": {
        "summary": "Remove Member",
        "description": "Removes a collaborator from a document. Owner-only.",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["memberId"],
                "properties": {
                  "memberId": { "type": "string", "format": "uuid", "example": "11111111-1111-4111-8111-111111111111" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collaborator successfully removed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Collaborator successfully removed." }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "403": { "$ref": "#/components/responses/403Forbidden" },
          "422": { "$ref": "#/components/responses/422ValidationError" }
        }
      }
    },
...
    "/api/documents/{id}/audit": {
      "get": {
        "summary": "Get Document Audit Log",
        "description": "Retrieves the full audit log history of document creation, deletion, role modification, snapshotting, and synchronization errors. Owner-only.",
        "parameters": [
          { "$ref": "#/components/parameters/id" },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "description": "Filter audit logs by specific AuditAction code",
            "schema": {
              "type": "string",
              "enum": ["DOCUMENT_CREATED", "DOCUMENT_DELETED", "ROLE_CHANGED", "SNAPSHOT_CREATED", "RESTORE_PERFORMED", "SYNC_FAILED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved audit logs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "action": { "type": "string", "enum": ["DOCUMENT_CREATED", "DOCUMENT_DELETED", "ROLE_CHANGED", "SNAPSHOT_CREATED", "RESTORE_PERFORMED", "SYNC_FAILED"] },
                          "userId": { "type": "string", "format": "uuid" },
                          "documentId": { "type": "string", "format": "uuid" },
                          "metadata": { "type": "object", "nullable": true },
                          "createdAt": { "type": "string", "format": "date-time" },
                          "user": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string", "example": "Alice" },
                              "email": { "type": "string", "example": "alice@example.com" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "403": { "$ref": "#/components/responses/403Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The unique UUID of the document",
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "401Unauthorized": {
        "description": "Request has expired token or is unauthenticated.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "UNAUTHORIZED" },
                    "message": { "type": "string", "example": "You must be signed in to access this resource." }
                  }
                }
              }
            }
          }
        }
      },
      "403Forbidden": {
        "description": "User has read access but does not possess required role permissions (e.g. editor or owner).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "FORBIDDEN" },
                    "message": { "type": "string", "example": "You do not have the required access permissions on this document." }
                  }
                }
              }
            }
          }
        }
      },
      "404NotFound": {
        "description": "Target document does not exist, or current user is not a member (preventing ID exposure).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "NOT_FOUND" },
                    "message": { "type": "string", "example": "Document not found" }
                  }
                }
              }
            }
          }
        }
      },
      "422ValidationError": {
        "description": "Zod Schema validation failed for request payload.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "VALIDATION_ERROR" },
                    "message": { "type": "string", "example": "Input validation failed." },
                    "details": {
                      "type": "object",
                      "properties": {
                        "fieldErrors": {
                          "type": "object",
                          "properties": {
                            "title": {
                              "type": "array",
                              "items": { "type": "string" },
                              "example": ["Title must be at least 1 character long."]
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}